# Conflicts:
#	frontend/javascripts/admin/dataset/remote/add_remote_layer.tsx
#	frontend/javascripts/admin/team/team_member_row.tsx
#	frontend/javascripts/components/pricing_enforcers.tsx
#	package.json

+++ webknossos-datastore/app/com/scalableminds/webknossos/datastore/controllers/BinaryDataController.scala
@@ -242,7 +242,7 @@ class BinaryDataController @Inject() (
             Some(dataSource.id),
             segmentationLayer,
             request.body.cuboid,
-            request.body.segmentId,
+            request.body.segmentId.toLong,
             request.body.voxelSizeFactorInUnit,
             tokenContextForRequest(using request),
             request.body.mapping,
